435 S 수정

+++ RC_DPU_PI_200M_4/src/rcfm_control.c
@@ -198,7 +198,7 @@ void SetRcfmStatBitFreq(const uint64_t Freq){
 	const uint8_t 	Prescaler = 2U;
 	const uint32_t	RefCLK = 50000000U;
 	uint64_t	TargetFreq = 0;
-	double		VCOCLK = 0;
+	double		VCOCLK = (double)0;
 	typTableLMX2582 ParamLMX2582 = {0,};
 
 
@@ -221,13 +221,13 @@ void SetRcfmStatBitFreq(const uint64_t Freq){
 
 
 		Division = (Segment[0] * Segment[1] * Segment[2]);
-		VCOCLK = ((double)(TargetFreq * (uint64_t)Division) / RefCLK);
+		VCOCLK = (double)((double)(TargetFreq * (uint64_t)Division) / RefCLK);
 
 		SendBuf = (0x0019U |(((uint16_t)SegValue[1] & 0x0FU) << 9)|(((uint16_t)SegEN[2] & 0x01U) << 8)|(((uint16_t)SegEN[1] & 0x01U) << 7)|(((uint16_t)SegValue[0] & 0x01U) << 2)|(((uint16_t)SegEN[0] & 0x01U) << 1));
 		SPI_WriteReg(LMX2592, 0x23U, SendBuf, 3U);
 
-		Int_Value  = (uint32_t)(VCOCLK / Prescaler);
-		Frac_Value = (uint16_t)(((VCOCLK / Prescaler)*1000U) - (Int_Value * 1000U));
+		Int_Value  = (uint32_t)((uint64_t)VCOCLK / Prescaler);
+		Frac_Value = (uint16_t)((((uint64_t)VCOCLK / Prescaler)*1000U) - (Int_Value * 1000U));
 
 
 		SendBuf = ((Int_Value & 0xFFFFU) * Prescaler);
